requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
--trusted-host=pip.astronomer.io
apache-airflow-providers-microsoft-mssql